Move unit tests out for league_pitching_stats
